requirements.txt
Django~=4.2.16
djangorestframework~=3.15.2
names~=0.3.0